Another fixes

+++ src/main/java/ru/avem/posum/models/SignalModel.java
@@ -264,7 +264,7 @@ public class SignalModel {
         this.calibrationExists = calibrationExists;
     }
 
-    public void setLoadsCounter(double loadsCounter) {
+    public void setLoadsCounter(int loadsCounter) {
         this.loadsCounter = loadsCounter;
         signalParametersModel.setLoadsCounter(loadsCounter);
     }
